mockplatform/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
dataclasses-json==0.6.7
decorator==5.1.1
Django==3.2.18
django-apscheduler==0.6.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-ckeditor==6.3.2
django-elasticsearch-dsl==8.0
django-js-asset==2.2.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
docx2txt==0.8
drf-yasg==1.20.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imgkit==1.2.3
inflection==0.5.1
interchange==2021.0.4
ipython==8.28.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jq==1.8.0
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.10.0
neo4j==5.25.0
networkx==3.4.2
numpy==1.26.4
ollama==0.3.3
openapi-codec==1.3.2
openpyxl==3.1.2
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pansi==2020.7.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prettytable==3.12.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
py2neo==2021.2.4
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
pyecharts==2.0.7
Pygments==2.18.0
PyMySQL==1.1.1
pypdf==5.0.1
pypinyin==0.53.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
simplejson==3.19.3
simpleui==4.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.39.0
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
watchdog==5.0.3
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.16.0
